Create Code->RockPaperScissors
I create three constants for our "Rock", "Scissors" and "Paper".  The program then prompts the player to choose one of them.  The program reads the player's choice and, through a variable of type Random, the computer selects a random number, which is used through a switch-case constitution to mark a random item from the given ones.  Finally, the win, loss and draw options are compared.